requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
boto3==1.28.84
botocore==1.31.84
certifi==2024.2.2
colorama==0.4.6
crispy-tailwind==0.5.0
distro==1.9.0
Django==4.2
django-browser-reload==1.12.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-htmx==1.17.0
django-phonenumber-field==7.2.0
django-storages==1.14.2
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
executing==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
icecream==2.1.3
idna==3.6
jmespath==1.0.1
numpy==1.26.4
openai==1.14.1
packaging==24.0
pandas==2.0.3
phonenumbers==8.13.23
Pillow==10.1.0
plotly==5.18.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
s3transfer==0.7.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
whitenoise==6.2.0
